Changes requested by reviewer, cleaned the code flow, refactored
