--- src/main/java/school/faang/user_service/service/util/AvatarApiService.java
@@ -1,26 +1,0 @@
-package school.faang.user_service.service.util;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestClientException;
-import org.springframework.web.client.RestTemplate;
-
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class AvatarApiService {
-    private final RestTemplate restTemplate;
-
-    public Optional<byte[]> getDefaultAvatar(String username) {
-        String url = "https://api.dicebear.com/9.x/initials/svg" +
-                "scale=50&backgroundType=gradientLinear&radius=50&seed=" + username;
-
-        try {
-            byte[] data = restTemplate.getForObject(url, byte[].class);
-            return Optional.ofNullable(data);
-        } catch(RestClientException e) {
-            return Optional.empty();
-        }
-    }
-}
